Develop

+++ app/src/main/java/com/arcathoria/character/dto/CharacterDTO.java
@@ -2,5 +2,5 @@ package com.arcathoria.character.dto;
 
 import java.util.UUID;
 
-public record CharacterDTO(UUID id, String characterName, Double health, Integer intelligence) {
+public record CharacterDTO(UUID id, String characterName, Integer health, Integer intelligence) {
 }
